4.6 creerFilm OK

+++ src/main/java/fr/eni/tp/filmotheque/bll/GenreService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface GenreService {
     List<Genre> findAllGenres();
-    Genre findGenreById(Long id);
+    Genre findGenreById(int id);
     Genre saveGenre(Genre genre);
 }
