+++ src/main/java/fr/eni/tp/filmotheque/bo/Personne.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 
 public abstract class Personne implements Serializable {
     private static final long serialVersionUID = 1L;
-    private long id;
+    private int id;
     private String nom;
     private String prenom;
 
@@ -16,7 +16,7 @@ public abstract class Personne implements Serializable {
         this.nom = nom;
         this.prenom = prenom;
     }
-    public Personne(long id, String nom, String prenom) {
+    public Personne(int id, String nom, String prenom) {
         this.id = id;
         this.nom = nom;
         this.prenom = prenom;
@@ -44,11 +44,11 @@ public abstract class Personne implements Serializable {
     }
 
     //getters and setters
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
